Add Payments in Buy Products API

+++ UserService/src/main/java/com/ecommerce/demo/service/BuyProductService.java
@@ -1,17 +1,16 @@
 package com.ecommerce.demo.service;
 
 import com.ecommerce.demo.dto.BuyProductDto;
-import com.ecommerce.demo.entity.CartItem;
-import com.ecommerce.demo.entity.ProductInventory;
+import com.ecommerce.demo.entity.*;
 import com.ecommerce.demo.exception.UserNotFoundException;
-import com.ecommerce.demo.repository.CartItemRepository;
-import com.ecommerce.demo.repository.ProductInventoryRepository;
-import com.ecommerce.demo.repository.ProductRepository;
-import com.ecommerce.demo.repository.UserRepository;
+import com.ecommerce.demo.repository.*;
+import jakarta.persistence.criteria.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BuyProductService {
@@ -21,30 +20,73 @@ public class BuyProductService {
 
     @Autowired
     CartItemRepository cartItemRepository;
+    @Autowired
+    UserRepository userRepository;
+    @Autowired
+    UserPaymentRepository userPaymentRepository;
 
-    public String buyProducts(int id,BuyProductDto buyProductDto) {
-        List<CartItem> cartItem1=cartItemRepository.findCartInUser(id);
-        for(CartItem cartItem:cartItem1)
-        {
-            int cartItemId=cartItem.getId();
-            for (CartItem cartItem2 : buyProductDto.getCartItemSet())
-            {
-                 int cartItemId2=cartItem2.getId();
-                if (cartItemId==cartItemId2){
-                    CartItem cartItem3 =cartItemRepository.findById(cartItemId).orElseThrow(() -> new IllegalArgumentException("CardId Not Found"));
-                    ProductInventory productInventory = cartItem3.getProduct().getProductInventory();
-                    int productQuantity = productInventory.getQuantity() - cartItem3.getQuantity();
-                    if(productQuantity <0)
-                    {
-                        throw new IllegalArgumentException("Product Inventory Has No Sufficient Quantity");
+    @Autowired
+    OrderDetailsRepository orderDetailsRepository;
+    @Autowired
+    OrderItemsRepository orderItemsRepository;
+
+
+    public String buyProducts(int id, BuyProductDto buyProductDto) {
+
+        Optional<Users> user = userRepository.findById(id);
+        if (user.isPresent()) {
+
+            OrderDetails orderDetails = new OrderDetails();
+
+            List<OrderItems> orderItemsList = new ArrayList<>();
+
+            UserPayment userPaymentRequestBody = buyProductDto.getUserPayment();
+            Users users = userRepository.findByUserPaymentsIdAndId(userPaymentRequestBody.getId(), id);
+
+            if (users != null) {
+                for (UserPayment userPayment : users.getUserPayments())
+                    if (userPayment.getId() == userPaymentRequestBody.getId()) {
+                        orderDetails.setUserPayments(userPayment);
+                        orderDetails.setUser(users);
                     }
-                     productInventory.setQuantity(productQuantity);
-                    productInventoryRepository.save(productInventory);
-                    cartItemRepository.deleteById(cartItemId);
-                    break;
+
+            } else {
+                throw new IllegalArgumentException("User Payment Is Not Exist");
+            }
+
+            List<CartItem> cartItem1 = cartItemRepository.findCartInUser(id);
+            for (CartItem cartItem : cartItem1) {
+                int cartItemId = cartItem.getId();
+                for (CartItem cartItem2 : buyProductDto.getCartItemSet()) {
+                    int cartItemId2 = cartItem2.getId();
+                    if (cartItemId == cartItemId2) {
+                        CartItem cartItem3 = cartItemRepository.findById(cartItemId).orElseThrow(() -> new IllegalArgumentException("CardId Not Found"));
+                        ProductInventory productInventory = cartItem3.getProduct().getProductInventory();
+                        int productQuantity = productInventory.getQuantity() - cartItem3.getQuantity();
+                        if (productQuantity < 0) {
+                            throw new IllegalArgumentException("Product Inventory Has No Sufficient Quantity");
+                        }
+                        productInventory.setQuantity(productQuantity);
+
+                        Product product = cartItem3.getProduct();
+
+                        OrderItems orderItems = new OrderItems();
+                        orderItems.setProduct(product);
+                        orderItemsList.add(orderItems);
+
+                        productInventoryRepository.save(productInventory);
+                        cartItemRepository.deleteById(cartItemId);
+                        break;
+                    }
+
                 }
             }
+            orderDetails.setOrderItems(orderItemsList);
+            orderDetailsRepository.save(orderDetails);
+        } else {
+            throw new UserNotFoundException("User Not Found");
         }
-        return  "Buy Products Succesfully";
+        return "Buy Products Successfully";
     }
+
 }
